+++ engine/spec/components/citizens_advice_components/checkbox_group_spec.rb
@@ -145,6 +145,39 @@ RSpec.describe CitizensAdviceComponents::CheckboxGroup, type: :component do
     it { is_expected.to have_css "[data-additional=example]" }
   end
 
+  context "when the page heading option is passed" do
+    before do
+      render_inline described_class.new(
+        legend: "Checkbox group field",
+        name: "checkboxes",
+        options: { page_heading: true }
+      ) do |c|
+        c.with_inputs(sample_inputs)
+      end
+    end
+
+    it { is_expected.to have_css "legend h1.cads-page-title", text: "Checkbox group field" }
+  end
+
+  context "when deprecated legend_heading option is provided" do
+    before do
+      allow(CitizensAdviceComponents.deprecator).to receive(:warn)
+
+      render_inline described_class.new(
+        legend: "Checkbox group field",
+        name: "checkboxes",
+        options: { legend_heading: true }
+      ) do |c|
+        c.with_inputs(sample_inputs)
+      end
+    end
+
+    it "logs deprecation warning" do
+      expect(CitizensAdviceComponents.deprecator).to have_received(:warn)
+        .with(/legend_heading option is deprecated/)
+    end
+  end
+
   private
 
   def sample_inputs